Integrate SCP driver into the platform driver
This patch merges the SCP driver into the platform driver, which clears
the way towards accessing the PMIC through the SCP instead of driving
the RSB bus directly by the platform driver. This way, the SCP can act
as the arbiter for accessing the PMIC by both the SCP firmware and
Genode components using the SCP service interface.