power/pinephone: detect power-button state
This patch adds support for sensing the power button. The power button
is attached to the PMIC, which is accessed via the SCP. To detect state
changes, the PMIC's IRQ (routed through the R_INTC to the GIC) is now
handled by the power driver. This has the added benefit that also other
interesting PMIC events (like connecting AC) get immediately reported.

Issue #22