sculpt/phone: screensaver
This patch adds a screensaver that kicks in after ten seconds of user
inactivity in the administrative user interface. It also can be manually
activated by pressing the power button. While the screen is blanked, a
press of the power button enables the display again.

Under the hood, Sculpt completely removes the drivers for the display
and the touchscreen while the screen is blanked, which considerably
reduces the power draw. The system also switches the CPU to economic
mode while the screen is blanked. Here are some illustrative data
points:

Max brightness in performance mode: 2.8 W
Max brightness in economic mode:    2.6 W
Low brightness in economic mode:    1.7 W
Screensaver:                        1.1 W