phone_manager: add speaker configuration menu
The speaker options configure if the big speaker on the back of the
PinePhone is used in addition to the small speaker on the front at the
top.

The default setting, 'Ring', configures its usage in a way that is
useful for regular phone-calls where the big speaker is enabled for
playing the ring-tones but is muted during a call. 'On' and 'Off'
control the speaker explicitly. To prevent unexpected echo issues where
the microphone at the back of the PinePhone picks received audio the
'Off' setting should be used.