WIP gpu/lima: remove IO signals from request handling
Initial I/O handling in this driver was inherited from the entaviv
driver but with the latest changes, like the context handling, the
way it operates no longer depends on the occurrence of such signals.

Returning from the scheduler by being blocked in the kernel will
always result in a failed request. How this failure is interpreted
depends on the type of requests. There are only two requests,
WAIT and WAIT_SYNCOBJ that deal with explicit synchronization, were
we deliberately block the task and failing is anticipated. The client
will wait for a notification signal in this case, which is sent
when the task is eventually unblocked.

Issue genodelabs/genode#2777.

Issue #27.